requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
jinja2
python-multipart
openai
python-dotenv
cryptography
mangum